# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Post API Documentation",
    "description": "Documentation for my Express Post API with validation using express-validator",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Posts",
      "description": "Operations related to blog posts"
    },
    {
      "name": "Comments",
      "description": "Operations related to comments on blog posts"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/posts/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/posts/{postId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/comments/{postId}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "text": {
                  "example": "any"
                },
                "author": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/comments/{commentId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "PostInput": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Example Post Title"
        },
        "description": {
          "type": "string",
          "example": "Example post description"
        }
      }
    },
    "PostUpdateInput": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Updated Title (optional)"
        },
        "description": {
          "type": "string",
          "example": "Updated description (optional)"
        }
      }
    },
    "Post": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "65f..."
        },
        "title": {
          "type": "string",
          "example": "Post Title"
        },
        "description": {
          "type": "string",
          "example": "Post Description"
        },
        "date": {
          "type": "string",
          "example": "2024-03-15T12:00:00.000Z"
        },
        "__v": {
          "type": "number",
          "example": 0
        }
      }
    },
    "CommentInput": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "example": "This is a comment text"
        },
        "author": {
          "type": "string",
          "example": "Comment Author Name"
        }
      }
    },
    "CommentUpdateInput": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "example": "Updated comment text (optional)"
        },
        "author": {
          "type": "string",
          "example": "Updated Author Name (optional)"
        }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "example": "660..."
        },
        "text": {
          "type": "string",
          "example": "Comment Text"
        },
        "author": {
          "type": "string",
          "example": "Author Name"
        },
        "postId": {
          "type": "string",
          "example": "65f..."
        },
        "date": {
          "type": "string",
          "example": "2024-03-22T10:00:00.000Z"
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Error message string"
        }
      }
    },
    "ValidationError": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "msg": {
                "type": "string",
                "example": "Validation error message"
              },
              "param": {
                "type": "string",
                "example": "fieldName"
              },
              "location": {
                "type": "string",
                "example": "body"
              }
            }
          }
        }
      }
    },
    "NotFoundError": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Resource not found"
        }
      }
    }
  }
}